refactor: improve navbar consistency
