+++ app/[lang]/smdd2024/components/projects.tsx
@@ -6,7 +6,7 @@ import GridItemsContainer, {
   GridItemsTitle
 } from "@/app/components/grid-items";
 import data from "@/data.json";
-import { SupportedLanguage } from "@/locales/.generated/types";
+import type { SupportedLanguage } from "@/locales/.generated/types";
 
 import { data as pageData } from "../data";
 
@@ -16,13 +16,13 @@ type OpenSourceProjectsProps = {
 
 export function OpenSourceProjects({ lang }: OpenSourceProjectsProps) {
   return (
-    <div className="container mx-auto w-full max-w-7xl px-2 text-black sm:px-0">
+    <section>
       <GridItemsContainer>
         <GridItemsTitle title="Open Source" />
         <GridItems items={pageData.projects} lang={lang} />
         <GridItemsMoreButton href={data.github} text="and more..." />
       </GridItemsContainer>
       <Script src="https://buttons.github.io/buttons.js" />
-    </div>
+    </section>
   );
 }
