+++ app/[lang]/smdd2024/data.ts
